+++ wallet/api/admin_update_network.go
@@ -10,8 +10,9 @@ import (
 )
 
 type AdminUpdateNetworkParams struct {
-	Name string `json:"name"`
-	API  struct {
+	Name     string             `json:"name"`
+	Metadata []network.Metadata `json:"metadata"`
+	API      struct {
 		GRPCConfig struct {
 			Hosts   []string `json:"hosts"`
 			Retries uint64   `json:"retries"`
@@ -23,8 +24,12 @@ type AdminUpdateNetworkParams struct {
 			Hosts []string `json:"hosts"`
 		} `json:"graphQLConfig"`
 	} `json:"api"`
+	Apps struct {
+		Console   string `json:"console"`
+		TokenDApp string `json:"tokenDApp"`
+		Explorer  string `json:"explorer"`
+	} `json:"apps"`
 }
-
 type AdminUpdateNetwork struct {
 	networkStore NetworkStore
 }
@@ -62,7 +67,8 @@ func validateUpdateNetworkParams(rawParams jsonrpc.Params) (network.Network, err
 	}
 
 	return network.Network{
-		Name: params.Name,
+		Name:     params.Name,
+		Metadata: params.Metadata,
 		API: network.APIConfig{
 			GRPC: network.GRPCConfig{
 				Hosts:   params.API.GRPCConfig.Hosts,
@@ -74,6 +80,11 @@ func validateUpdateNetworkParams(rawParams jsonrpc.Params) (network.Network, err
 			GraphQL: network.GraphQLConfig{
 				Hosts: params.API.GraphQLConfig.Hosts,
 			},
+		},
+		Apps: network.AppsConfig{
+			Console:   params.Apps.Console,
+			TokenDApp: params.Apps.TokenDApp,
+			Explorer:  params.Apps.Explorer,
 		},
 	}, nil
 }