+++ wallet/network/store/v1/file_store.go
@@ -11,6 +11,7 @@ import (
 
 	vgfs "code.vegaprotocol.io/vega/libs/fs"
 	"code.vegaprotocol.io/vega/paths"
+	"code.vegaprotocol.io/vega/wallet/api"
 
 	"code.vegaprotocol.io/vega/wallet/network"
 )
@@ -23,7 +24,7 @@ var (
 	ErrNetworkNameCannotContainSlash = errors.New("the network name cannot contain `\\` nor `/`")
 )
 
-type Store struct {
+type FileStore struct {
 	networksHome string
 }
 
@@ -31,18 +32,18 @@ type networkFileContent struct {
 	API network.APIConfig `json:"api"`
 }
 
-func InitialiseStore(vegaPaths paths.Paths) (*Store, error) {
+func InitialiseStore(vegaPaths paths.Paths) (*FileStore, error) {
 	networksHome, err := vegaPaths.CreateConfigDirFor(paths.WalletServiceNetworksConfigHome)
 	if err != nil {
 		return nil, fmt.Errorf("couldn't get config path for %s: %w", paths.WalletServiceNetworksConfigHome, err)
 	}
 
-	return &Store{
+	return &FileStore{
 		networksHome: networksHome,
 	}, nil
 }
 
-func (s *Store) ListNetworks() ([]string, error) {
+func (s *FileStore) ListNetworks() ([]string, error) {
 	networksParentDir, networksDir := filepath.Split(s.networksHome)
 	entries, err := fs.ReadDir(os.DirFS(networksParentDir), networksDir)
 	if err != nil {
@@ -62,15 +63,15 @@ func (s *Store) ListNetworks() ([]string, error) {
 	return networks, nil
 }
 
-func (s *Store) GetNetworksPath() string {
+func (s *FileStore) GetNetworksPath() string {
 	return s.networksHome
 }
 
-func (s *Store) GetNetworkPath(name string) string {
+func (s *FileStore) GetNetworkPath(name string) string {
 	return s.nameToFilePath(name)
 }
 
-func (s *Store) NetworkExists(name string) (bool, error) {
+func (s *FileStore) NetworkExists(name string) (bool, error) {
 	if err := ensureValidNetworkName(name); err != nil {
 		return false, err
 	}
@@ -78,7 +79,7 @@ func (s *Store) NetworkExists(name string) (bool, error) {
 	return vgfs.FileExists(s.GetNetworkPath(name))
 }
 
-func (s *Store) GetNetwork(name string) (*network.Network, error) {
+func (s *FileStore) GetNetwork(name string) (*network.Network, error) {
 	if err := ensureValidNetworkName(name); err != nil {
 		return nil, err
 	}
@@ -93,7 +94,7 @@ func (s *Store) GetNetwork(name string) (*network.Network, error) {
 	}, nil
 }
 
-func (s *Store) SaveNetwork(net *network.Network) error {
+func (s *FileStore) SaveNetwork(net *network.Network) error {
 	if err := ensureValidNetworkName(net.Name); err != nil {
 		return err
 	}
@@ -107,7 +108,7 @@ func (s *Store) SaveNetwork(net *network.Network) error {
 	return nil
 }
 
-func (s *Store) DeleteNetwork(name string) error {
+func (s *FileStore) DeleteNetwork(name string) error {
 	if err := ensureValidNetworkName(name); err != nil {
 		return err
 	}
@@ -115,11 +116,37 @@ func (s *Store) DeleteNetwork(name string) error {
 	return os.Remove(path)
 }
 
-func (s *Store) nameToFilePath(network string) string {
+func (s *FileStore) RenameNetwork(currentName, newName string) error {
+	if err := ensureValidNetworkName(currentName); err != nil {
+		return err
+	}
+
+	if err := ensureValidNetworkName(newName); err != nil {
+		return err
+	}
+
+	currentNetworkPath := s.nameToFilePath(currentName)
+
+	if exists, err := vgfs.PathExists(currentNetworkPath); err != nil {
+		return fmt.Errorf("could not verify the path at %s: %w", currentNetworkPath, err)
+	} else if !exists {
+		return api.ErrNetworkDoesNotExist
+	}
+
+	newNetworkPath := s.nameToFilePath(newName)
+
+	if err := os.Rename(currentNetworkPath, newNetworkPath); err != nil {
+		return fmt.Errorf("could not rename the network %q to %q at %s: %w", currentName, newName, s.networksHome, err)
+	}
+
+	return nil
+}
+
+func (s *FileStore) nameToFilePath(network string) string {
 	return filepath.Join(s.networksHome, network+fileExt)
 }
 
-func (s *Store) fileNameToName(fileName string) string {
+func (s *FileStore) fileNameToName(fileName string) string {
 	return fileName[:len(fileName)-len(fileExt)]
 }